Now Compatible with Coq8.5~beta3. This fixes bug #3.
Merging constraints into a constraint graph now first adds all the necessary universe levels to the graph as is required as of Coq8.5~beta3.
Test suit is updated to reflect the new behavior due to Coq8.5~beta3.
This version is not backwards compatible to Coq8.5~beta2 or earlier. This is because of changes in the kernel API.